ai_bot_vocabulary/ai_vocab_bot-master/requirements.txt
aiogram==3.0.0b6
beautifulsoup4==4.12.3
gensim==4.3.3
googletrans==4.0.0rc1
numpy==1.26.4
psycopg2_binary==2.9.9
Requests==2.32.3
google-generativeai==0.8.1
